requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.41
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
python-dotenv==1.0.1
aiofiles==23.2.1
email-validator==2.1.1
psycopg[binary]==3.2.10
slowapi==0.1.9
apscheduler==3.10.4

# Testes
pytest==8.2.2
pytest-asyncio==0.23.7
httpx==0.27.0
cloudinary==1.44.2
